# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prevailing Wage Public API",
    "version": "1.0.0",
    "description": "Read-only REST API for accessing prevailing wage project, payroll, worker, and compliance data. All endpoints require Bearer API key authentication. Rate limited to 1000 req/hour per API key."
  },
  "servers": [
    {
      "url": "https://prevailingwage.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Invalid API key"
          }
        }
      },
      "Meta": {
        "type": "object",
        "required": ["requestId", "timestamp"],
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique request identifier for tracing"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 response timestamp"
          }
        }
      },
      "PaginatedMeta": {
        "allOf": [
          { "$ref": "#/components/schemas/Meta" },
          {
            "type": "object",
            "required": ["page", "limit", "total", "hasNext"],
            "properties": {
              "page": { "type": "integer", "minimum": 1 },
              "limit": { "type": "integer", "minimum": 1, "maximum": 100 },
              "total": { "type": "integer", "minimum": 0 },
              "hasNext": { "type": "boolean" }
            }
          }
        ]
      },
...
      "PayrollWeek": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "weekEndingDate": {
            "type": ["string", "null"],
            "description": "Week ending date (ISO 8601)"
          },
          "payrollNumber": {
            "type": ["integer", "null"],
            "description": "Sequential payroll number within the project"
          },
          "isFinal": {
            "type": ["boolean", "null"],
            "description": "Whether this is a final (certified) payroll"
          },
          "submittedAt": {
            "type": ["string", "null"],
            "description": "Submission timestamp (ISO 8601)"
          },
          "submittedTo": {
            "type": ["string", "null"],
            "description": "Name of agency or contact payroll was submitted to"
          },
          "createdAt": {
            "type": ["string", "null"],
            "description": "Record creation timestamp (ISO 8601)"
          }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based)",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Results per page (max 100)",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "ProjectIdParam": {
        "name": "id",
        "in": "path",
        "description": "Project UUID",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Invalid API key" }
          }
        }
      },
      "Forbidden": {
        "description": "API key lacks required scope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Scope required: projects:read" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded (1000 req/hour per API key)",
        "headers": {
          "X-RateLimit-Limit": { "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
          "X-RateLimit-Reset": { "schema": { "type": "integer", "description": "Unix timestamp when window resets" } },
          "X-RateLimit-Window": { "schema": { "type": "string", "example": "3600s" } }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Rate limit exceeded" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/projects/{id}/payroll-weeks": {
      "get": {
        "tags": ["Public API v1"],
        "summary": "List payroll weeks for a project",
        "description": "Returns a paginated list of payroll weeks for the specified project. Rate limited to 1000 req/hour per API key. Required scope: `payroll:read`.",
        "operationId": "listPayrollWeeks",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdParam" },
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of payroll weeks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/PayrollWeek" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginatedMeta" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "Project not found or not accessible",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  }
}